services/unitable/pyproject.toml
[project]
name = "unitable"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "robyn>=0.61.2",
    "apted==1.0.3",
    "beautifulsoup4==4.12.3",
    "distance==0.1.3",
    "einops==0.8.0",
    "faster-coco-eval==1.6.3",
    "hydra-colorlog==1.2.0",
    "hydra-core==1.3.2",
    "jsonlines==4.0.0",
    "lxml==4.9.3",
    "matplotlib==3.9.2",
    "ptflops==0.7.4",
    "pycocotools==2.0.8",
    "robyn==0.61.2",
    "tokenizers==0.20.1",
    "torch==2.2.1",
    "torchaudio==2.2.1",
    "torchdata==0.7.1",
    "torchmetrics==1.4.3",
    "torchtext==0.17.1",
    "torchvision==0.17.1",
    "wandb==0.18.3",
    "python-dotenv>=1.0.1",
    "opencv-python>=4.10.0.84",
]
